Replaced 'no' langauge code with 'nb' and use full norwegian language names
The 'no' language code is a obsolete language code that is the union
of language codes 'nb' (Norwegian Bokmål) and 'nn' (Norwegian Nynorsk),
the two written variants in use in Norway.  As 'no' is misleading, and
seem to be used by Whisper to mean Norwegian Bokmål, I recommend replacing
it with 'nb' and using the full names for both of the norwegian written
forms.

Useful references:

 * https://linguistics.stackexchange.com/questions/36784/norwegian-translation-codes-no-nn-nb-which-to-use-on-a-website
 * https://people.skolelinux.org/pere/blog/Spr_kkoder_for_POSIX_locale_i_Norge.html