Teach whisper to look for models in /usr/ and /var/ too.
This allow the model to be provided in OS packages and allow users to
share model files instead of downloading individual copies to each
users home directory.

Existing models in the users home directory is preferred, and files
in /var/ are preferred over files in /usr/.  If no model is found in
the global paths, the model is downloaded into ~/.cache/whisper/ as
before.